+++ noodles-fastq/examples/fastq_count_async.rs
@@ -16,7 +16,7 @@ async fn main() -> io::Result<()> {
     let mut reader = File::open(src)
         .await
         .map(BufReader::new)
-        .map(fastq::AsyncReader::new)?;
+        .map(fastq::r#async::io::Reader::new)?;
 
     let mut records = reader.records();
     let mut n = 0;