+++ noodles-fastq/src/async/io.rs
@@ -1,0 +1,5 @@
+//! Async FASTQ I/O.
+
+mod reader;
+
+pub use self::reader::Reader;