+++ noodles-fastq/src/io.rs
@@ -1,0 +1,5 @@
+//! FASTQ I/O.
+
+pub mod reader;
+
+pub use self::reader::Reader;